Mark comment threads read once they have been seen
A thread counts as read after staying in the middle of the viewport
long enough to read it, so scrolling past leaves it unread. Threads
hiding their replies are left alone until expanded.

Splits Thread-spec into topic specs under Thread/features, since a
flat spec file and a features directory for the same unit must not
coexist.
